first stage shoppin cart

+++ Back-end/src/main/java/nl/sander/service/AbstractCrudRepository.java
@@ -28,7 +28,7 @@ public abstract class AbstractCrudRepository<T> {
         this.entityManager.remove(attached);
     }
 
-    public T find(Long key) {
+    public T find(int key) {
         return this.entityManager.find(getEntityClass(), key);
     }
